Add files via upload
Here is one of my latest works.

This is a currency value calculator. In this form, it is possible to find out the exchange rate and the cost of the transfer in two currencies (in local currency and in dollars). Added an algorithm for calculating the size of the commission depending on the transfer amount.
Also it was necessary, in the form add the ability to enter different exchange rates for different transfer sizes (for amounts up to $1.000, for $2.000, etc.).

The form is used on the pages of MPS group websites.

Sites on Wordpress, the calculator is written on php, form content and styles on HTML and CSS and calculation logic on JS (JQuery).